Three story slides added to the dashboard
changed from mean to median based on mireks answer on email

"DTN is always calculated as median, not average. We even filter out values 400+ as improbable as it is way outside of the procedure time-window.
Value under 10 are questionable. As you have to enroll the patient, do CT, analyze CT and inject thrombolytic. It can be mistake or they simply didn't do CT and risked killing the patient etc."